start quiz helper + logging
